Two more TypeError fixes related to numpy 1.12

+++ urbansim/urbanchoice/tests/test_mnl.py
@@ -94,7 +94,7 @@ def choosers(test_data):
 @pytest.fixture
 def chosen(df, num_alts, test_data):
     return df[test_data['column']].values.astype('int').reshape(
-        (len(df) / num_alts, num_alts))
+        (int(len(df) / num_alts), num_alts))
 
 
 @pytest.fixture
